add keybinding to switch and zoom
This is intended for when you're running your tests in a second smaller
pane and want to take a quick peek at what the test output is.

You can now press `PREFIX J` to select the bottom pane and zoom it at
the same time. To escape from the zoomed view you simply press `PREFIX
k` to jump back up or simply `PREFIX j` because the direction doesn't
matter in a two pane setup.